1.答辩 2.管理员 3.优化

+++ src/main/java/com/jxufe/ljw/thesis/bean/ReplyScore.java
@@ -19,7 +19,7 @@ public class ReplyScore {
     private String replyPlace;
     private  int replyStatus;
     private String replyOpinion;
-    private double openReportScore;
+    private int openReportScore;
     private  int openReportStatus;
     private String openReportOpinion;
     private int reviewStatus;
