+++ src/main/java/com/jxufe/ljw/thesis/bean/StudentTeacherRelation.java
@@ -65,7 +65,7 @@ public class StudentTeacherRelation {
     private String replyPlace;
     private  int replyStatus;
     private String replyOpinion;
-    private double openReportScore;
+    private int openReportScore;
     private  int openReportStatus;
     private String openReportOpinion;
     private int reviewStatus;
